Generalized strToHex function, address tracker
I generalized the strToHex function so that it now takes in any hex number represented in string and spits out the correct hex/int number, signed and unsigned. I also added some functionality to parse each opcode and its parameter to a dictionary so that we can manage to add labels when we encounter any of the branch opcodes.