Branch and labels are functional now
I got branching with the branching opcodes to work by adding their _a suffix to them so that they can branch to a label now rather than their byte offsets. Now, I got to get JMP to work, but that's going to be hard since they jump to a specific address rather than an offset to an address.